[FIX] Fixing warning in mason_variator.cpp.

+++ extras/apps/mason2/mason_variator.cpp
@@ -290,7 +290,7 @@ public:
         for (unsigned i = 0; i < length(variationSizeRecords); ++i)
         {
             VariationSizeRecord const & record = variationSizeRecords[i];
-            if (variationToContig[i] != rId)
+            if (variationToContig[i] != (int)rId)
                 continue;  // This variation does not belong here.
             int const MAX_TRIES = 1000;
             int tries = 0;
